Feat/config inheritance (#126)
* feat: inheritance for datasource configs

Introduce datasource configuration inheritance to the lumi configuration.

This means that e.g. fdb config can now be defined in a single location as a "datasource" that is not initiated due ot not being included directly in any collections. Other datasources (fdb, polytope, mars) can inherit this datasource config and not have to define their own.

Further such deduplications can be made on an ad-hoc basis

BREAKING CHANGES:
- mars datasource fdb_config no longer a list
- fdb datasource config field now called fdb_config